+++ Api/Dto/ModelDtos/DirectiveDto.cs
@@ -8,7 +8,7 @@ namespace Api.Dto
         public int DirectiveId { get; set; }
         public string Name { get; set; }
         public int NumberOfExpenses { get; set; }
-        public int TotalCostOfExpenses { get; set; }
+        public decimal TotalCostOfExpenses { get; set; }
         public ICollection<ExpenseDto> Expenses { get; set; }
     }
 }
