+++ Api/Interfaces/IAnalyticService.cs
@@ -8,14 +8,13 @@ namespace Api.Interfaces
     public interface IAnalyticService
     {
         Task UpdateAnalytics(ApplicationUser currentser, Expense changedExpense);
-
         void CreateGeneralAnalytic(ApplicationUser currentser, Expense changedExpense);
         int MostExpensiveLookup(List<Expense> userExpenses, string lookupIdProperty, Expense changedExpense, ApplicationUser user);
         void UpdateLookupAnalyticFromScratch(List<Expense> userExpenses, string lookupIdProperty, ApplicationUser user, Expense changedExpense);
         List<LookupCount> BuildLookupCountCollection(List<Expense> sortedExpenses, string lookupIdProperty, ApplicationUser user);
-        LookupCount CreateLookupCount(int LookupId, int TotalCostOfExpenses, int NumberOfExpenses, ApplicationUser user, string lookupTypeName);
+        LookupCount CreateLookupCount(int LookupId, decimal TotalCostOfExpenses, int NumberOfExpenses, ApplicationUser user, string lookupTypeName);
         void UpdateLookCollectionForNoExpenses(LookupAnalytic existingLookupAnalytic);
         void UpdateLookupAnalytic(List<LookupCount> lookupCollection, DateTime date, ApplicationUser user, LookupAnalytic existingLookupAnalytic, string lookupTypeName);
-        int GetTotalCost(List<Expense> userExpenses);
+        decimal GetTotalCost(List<Expense> userExpenses);
     }
 }
